Complete commit
Hoàn thành chọn một sản phẩm không thông qua chi tiết sản phẩm vào giỏ hàng
Có chức năng Check out để lấy thông tin cần thiết cho hóa đơn và kiểm đăng nhập, đồng thời kiểm tra phương thức thanh toán.
Thanh toán bằng tiền mặt, thanh toán khi giao hàng, thanh toán paypal và có thể chuyển qua thanh toán thành công hay hủy thanh toán.
Có chỉnh sửa home controller để phục vụ việc thanh toán. Chỉnh sửa header một chút để có đường dẫn tới những trang cần thiết. Chỉnh sửa các lớp Entity Order, OrderDetail và các lớp service liên quan đến hai phần đó.
Thêm một package util có PaypalUtils và các lớp trong gói config gồm (PaypalConfig, PaypalPaymentIntent, PaypalPaymentMethod). Thêm một lớp DTO CartItem.

+++ WebProject/src/main/java/com/Controller/Client/ShoppingCart/ShoppingCartController.java
@@ -42,7 +42,7 @@ public class ShoppingCartController {
 	}
 	
 	@GetMapping("/Add/{id}")
-	public String addItem(@PathVariable("id")long id) {
+	public String addItem(@PathVariable("id")Long id) {
 		Product product = productService.getProductById(id);
 		CartItem item = new CartItem(product, 1);
 		shoppingCartService.addItem(item);
@@ -50,13 +50,13 @@ public class ShoppingCartController {
 	}
 	
 	@GetMapping("/Delete/{id}")
-	public String deleteItem(@PathVariable("id")long id) {
+	public String deleteItem(@PathVariable("id")Long id) {
 		shoppingCartService.removeItem(id);
 		return "redirect:/Client/ShoppingCart/Show";
 	}
 	
 	@PostMapping("/Update")
-	public String updateCart(@RequestParam("id") int id, @RequestParam("quantity")int quantity) {
+	public String updateCart(@RequestParam("id")Long id, @RequestParam("quantity")Integer quantity) {
 		shoppingCartService.updateItem(id, quantity);
 		return "redirect:/Client/ShoppingCart/Show";
 	}
